fix oop structural issues
