+++ Resources/python/lerobot_env_schola/src/lerobot_env_schola/__init__.py
@@ -3,12 +3,10 @@
 """Schola environment plugin for LeRobot."""
 
 from .config import ScholaEnvConfig, ScholaObservationConfig
-from .observations import ObservationAdapter
 from .vector_env import LeRobotScholaVectorEnv
 
 __all__ = [
     "LeRobotScholaVectorEnv",
-    "ObservationAdapter",
     "ScholaEnvConfig",
     "ScholaObservationConfig",
 ]
